+++ packages/pluggableWidgets/rich-text-web/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We disabled the version check from CKEditor 4, which generated security alert in all widget instances.
+
+### Added
+
+-   We add validation message for rich text editor.
+
 ## [2.2.4] - 2023-11-29
 
 ### Fixed